Add VS2019 solution and fix compilation errors/warnings
- Added scetool.sln, scetool.vcxproj, and scetool.vcxproj.filters.
- Fixed POSIX deprecation errors for strdup and access in types.h.
- Resolved _CRT_SECURE_NO_WARNINGS redefinition in getopt.c.
- Fixed format string and type conversion warnings in multiple files.
- Added .gitignore for build artifacts.
- Added README.VS.md with build instructions.

+++ frontend.cpp
@@ -54,7 +54,7 @@ static BOOL _is_hexdigit(s8 c)
 
 static BOOL _is_hexnumber(const s8 *str)
 {
-	u32 i, len = strlen(str);
+	u32 i, len = (u32)strlen(str);
 	for(i = 0; i < len; i++)
 		if(_is_hexdigit(str[i]) == FALSE)
 			return FALSE;
@@ -151,7 +151,7 @@ static BOOL _fill_self_config(self_config_t *sconf)
 		printf("[*] Error (Key Revision): Please provide a valid hexadecimal number.\n");
 		return FALSE;
 	}
-	sconf->key_revision = _x_to_u64(_key_rev);
+	sconf->key_revision = (u16)_x_to_u64(_key_rev);
 
 	if(_auth_id == NULL)
 	{
@@ -165,7 +165,7 @@ static BOOL _fill_self_config(self_config_t *sconf)
 		printf("[*] Error: Please specify a vendor ID.\n");
 		return FALSE;
 	}
-	sconf->vendor_id = _x_to_u64(_vendor_id);
+	sconf->vendor_id = (u32)_x_to_u64(_vendor_id);
 
 	if(_self_type == NULL)
 	{
@@ -178,7 +178,7 @@ static BOOL _fill_self_config(self_config_t *sconf)
 		printf("[*] Error: Invalid SELF type.\n");
 		return FALSE;
 	}
-	sconf->self_type = type;
+	sconf->self_type = (u32)type;
 
 	if(_app_version == NULL)
 	{
@@ -275,7 +275,7 @@ static BOOL _fill_npdrm_config(self_config_t *sconf)
 		printf("[*] Error: Invalid application type.\n");
 		return FALSE;
 	}
-	sconf->npdrm_config->app_type = type;
+	sconf->npdrm_config->app_type = (u32)type;
 
 	if(_content_id == NULL)
 	{
